POTUS Analysis
What do presidents of the United States have in common with each other? In this project, a dataset is created by scraping multiple webpages using Python's Requests and BeautifulSoup libraries. After the pandas dataframe is created and cleaned, the data is visualized using the Seaborn library. Tableau is also used to create an interactive dashboard that is more readable to people with non-technical backgrounds.